requirements.txt
aiogram>=3.7.0
SQLAlchemy>=2.0.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
fake-useragent>=1.4.0
python-dotenv>=1.0.0
undetected-chromedriver>=3.0.0
selenium>=4.0.0